revert back and invalidate cache
