--- src/test/e2e/36_health_check_crd_test.go
@@ -1,34 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-FileCopyrightText: 2021-Present The Zarf Authors
-
-// Package test provides e2e tests for Zarf.
-package test
-
-import (
-	"fmt"
-	"path/filepath"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
-
-// TestHealthCheckCRDRegisteredMidDeploy guards against a regression where healthchecks
-// don't recognize and wait against a crd that was introduced during package deploy
-func TestHealthCheckCRDRegisteredMidDeploy(t *testing.T) {
-	t.Log("E2E: Health check for a CRD registered mid-deploy")
-
-	tmpdir := t.TempDir()
-	_, _, err := e2e.Zarf(t, "package", "create", "src/test/packages/36-health-check-crd", "-o", tmpdir, "--confirm")
-	require.NoError(t, err)
-
-	packageName := fmt.Sprintf("zarf-package-health-check-crd-%s.tar.zst", e2e.Arch)
-	path := filepath.Join(tmpdir, packageName)
-
-	defer func() {
-		_, _, err := e2e.Zarf(t, "package", "remove", "health-check-crd", "--confirm")
-		require.NoError(t, err)
-	}()
-
-	_, _, err = e2e.Zarf(t, "package", "deploy", path, "--timeout", "3m", "--confirm")
-	require.NoError(t, err)
-}
